Refatoração das abas, dividindo por UserControls.
Limpeza do código.
