+++ MaxTool/Modules/Install.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Drawing;
+using System.Data;
+using System.Linq;
+using System.Text;
+using System.Windows.Forms;
+
+namespace MaxTools.Modules
+{
+    public partial class Install : UserControl
+    {
+        public Install()
+        {
+            InitializeComponent();
+        }
+    }
+}
